+++ .gitignore
@@ -7,3 +7,8 @@ Control-MM/hardware/__pycache__/APump.cpython-38.pyc
 Control-MM/fluidics/__pycache__/FluidicsProgramBuilder.cpython-38.pyc
 Control-MM/fluidics/__pycache__/FluidicsProgramBuilder.cpython-38.pyc
 Control-MM/fluidics/__pycache__/FluidicsProgramBuilder.cpython-38.pyc
+*.pyc
+Control-MM/.vscode/settings.json
+Control-MM/.vscode/settings.json
+*.csv
+control/.vscode/settings.json
