--- Control-MM/.vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "python.pythonPath": "C:\\Users\\dpshe\\miniconda3\\envs\\pyqt\\python.exe",
-    "jupyter.jupyterServerType": "local"
-}
