--- Control-MM/20210208_firstround.csv
@@ -1,7 +1,0 @@
-,round,source,volume,time
-0,1,SSC,2.0,5.0
-1,1,B01,2.0,5.0
-2,1,B01,2.0,6.0
-3,1,PAUSE,5.0,15.0
-4,1,READOUT WASH,2.0,9.0
-5,1,IMAGING BUFFER,2.0,6.0
