--- Control-MM/test_load_new_dataset.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-
-'''
-OPM stage control with fluidics
-
-Shepherd 02/21
-'''
-
-# imports
-from pycromanager import Dataset
-
-dataset = Dataset('E:\\test_20210209\\test_4')
